Improved slither launcher; now lets us workaround specific slither-crashing input files.
On invocation, facadeAct.sol is moved out of the way. It is moved back
even if the user breaks out of the slither invocation early.

On approaching this flexibility, I decided that our slither wrapper had
outgrown the comfortable level of complexity for being a bash script,
and it was worth brushing off python scripting stuff. So:

- tools/slither has been replaced by tools/slither.py
- "yarn slither" invocation uses tools/slither.py